strip non-essential data from redis cache txs
